+++ somni-parser/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Parser no longer needs the whole source to be tokenized upfront
 - Added TypeSet type parameter to the whole AST and the `parse` function
 - `Expression::as_variable` has been removed
+- Split `Expression` and `RightHandExpression`, added `LeftHandExpression`
 
 # [0.1.0] - 2025-07-24